requirements.txt
requests
feedgen
python-pytz 